+++ CHANGELOG.md
@@ -35,6 +35,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - Category is extended
     - Administrator is extended
     - Transport is extended
+    - Product is extended
 
 ### [shopsys/coding-standards]
 #### Added